+++ common/src/main/java/com/yanny/ali/plugin/Plugin.java
@@ -2,8 +2,8 @@ package com.yanny.ali.plugin;
 
 import com.yanny.ali.api.*;
 import com.yanny.ali.plugin.client.widget.*;
-import com.yanny.ali.plugin.client.widget.trades.EmeraldForVillagerTypeItemWidget;
 import com.yanny.ali.plugin.client.widget.trades.ItemListingWidget;
+import com.yanny.ali.plugin.client.widget.trades.SubTradesWidget;
 import com.yanny.ali.plugin.client.widget.trades.TradeLevelWidget;
 import com.yanny.ali.plugin.client.widget.trades.TradeWidget;
 import com.yanny.ali.plugin.common.EntityUtils;
@@ -46,16 +46,7 @@ public class Plugin implements IPlugin {
         registry.registerWidget(TradeNode.ID, TradeWidget::new);
         registry.registerWidget(TradeLevelNode.ID, TradeLevelWidget::new);
         registry.registerWidget(ItemStackNode.ID, ItemStackWidget::new);
-        registry.registerWidget(EmeraldForItemsNode.ID, ItemListingWidget::new);
-        registry.registerWidget(ItemsForEmeraldsNode.ID, ItemListingWidget::new);
-        registry.registerWidget(ItemsAndEmeraldsToItemsNode.ID, ItemListingWidget::new);
-        registry.registerWidget(SuspiciousStewForEmeraldNode.ID, ItemListingWidget::new);
-        registry.registerWidget(EnchantedItemForEmeraldsNode.ID, ItemListingWidget::new);
-        registry.registerWidget(TippedArrowForItemsAndEmeraldsNode.ID, ItemListingWidget::new);
-        registry.registerWidget(TreasureMapForEmeraldsNode.ID, ItemListingWidget::new);
-        registry.registerWidget(EnchantBookForEmeraldsNode.ID, ItemListingWidget::new);
-        registry.registerWidget(DyedArmorForEmeraldsNode.ID, ItemListingWidget::new);
-        registry.registerWidget(EmeraldsForVillagerTypeItemNode.ID, EmeraldForVillagerTypeItemWidget::new);
+        registry.registerWidget(SubTradesNode.ID, SubTradesWidget::new);
         registry.registerWidget(ItemsToItemsNode.ID, ItemListingWidget::new);
 
         registry.registerNode(LootTableNode.ID, LootTableNode::new);
@@ -73,16 +64,7 @@ public class Plugin implements IPlugin {
         registry.registerNode(TradeNode.ID, TradeNode::new);
         registry.registerNode(TradeLevelNode.ID, TradeLevelNode::new);
         registry.registerNode(ItemStackNode.ID, ItemStackNode::new);
-        registry.registerNode(EmeraldForItemsNode.ID, EmeraldForItemsNode::new);
-        registry.registerNode(ItemsForEmeraldsNode.ID, ItemsForEmeraldsNode::new);
-        registry.registerNode(ItemsAndEmeraldsToItemsNode.ID, ItemsAndEmeraldsToItemsNode::new);
-        registry.registerNode(SuspiciousStewForEmeraldNode.ID, SuspiciousStewForEmeraldNode::new);
-        registry.registerNode(EnchantedItemForEmeraldsNode.ID, EnchantedItemForEmeraldsNode::new);
-        registry.registerNode(TippedArrowForItemsAndEmeraldsNode.ID, TippedArrowForItemsAndEmeraldsNode::new);
-        registry.registerNode(TreasureMapForEmeraldsNode.ID, TreasureMapForEmeraldsNode::new);
-        registry.registerNode(EnchantBookForEmeraldsNode.ID, EnchantBookForEmeraldsNode::new);
-        registry.registerNode(DyedArmorForEmeraldsNode.ID, DyedArmorForEmeraldsNode::new);
-        registry.registerNode(EmeraldsForVillagerTypeItemNode.ID, EmeraldsForVillagerTypeItemNode::new);
+        registry.registerNode(SubTradesNode.ID, SubTradesNode::new);
         registry.registerNode(ItemsToItemsNode.ID, ItemsToItemsNode::new);
     }
 
@@ -178,27 +160,27 @@ public class Plugin implements IPlugin {
         registry.registerItemStackModifier(SetNbtFunction.class, TooltipUtils::applyItemStackModifier);
         registry.registerItemStackModifier(SetPotionFunction.class, TooltipUtils::applyItemStackModifier);
 
-        registry.registerItemListing(VillagerTrades.DyedArmorForEmeralds.class, DyedArmorForEmeraldsNode::new);
-        registry.registerItemListing(VillagerTrades.EmeraldForItems.class, EmeraldForItemsNode::new);
+        registry.registerItemListing(VillagerTrades.DyedArmorForEmeralds.class, TradeUtils::getNode);
+        registry.registerItemListing(VillagerTrades.EmeraldForItems.class, TradeUtils::getNode);
         registry.registerItemListing(VillagerTrades.EmeraldsForVillagerTypeItem.class, EmeraldsForVillagerTypeItemNode::new);
-        registry.registerItemListing(VillagerTrades.EnchantBookForEmeralds.class, EnchantBookForEmeraldsNode::new);
-        registry.registerItemListing(VillagerTrades.EnchantedItemForEmeralds.class, EnchantedItemForEmeraldsNode::new);
-        registry.registerItemListing(VillagerTrades.ItemsAndEmeraldsToItems.class, ItemsAndEmeraldsToItemsNode::new);
-        registry.registerItemListing(VillagerTrades.ItemsForEmeralds.class, ItemsForEmeraldsNode::new);
-        registry.registerItemListing(VillagerTrades.SuspiciousStewForEmerald.class, SuspiciousStewForEmeraldNode::new);
-        registry.registerItemListing(VillagerTrades.TippedArrowForItemsAndEmeralds.class, TippedArrowForItemsAndEmeraldsNode::new);
-        registry.registerItemListing(VillagerTrades.TreasureMapForEmeralds.class, TreasureMapForEmeraldsNode::new);
+        registry.registerItemListing(VillagerTrades.EnchantBookForEmeralds.class, TradeUtils::getNode);
+        registry.registerItemListing(VillagerTrades.EnchantedItemForEmeralds.class, TradeUtils::getNode);
+        registry.registerItemListing(VillagerTrades.ItemsAndEmeraldsToItems.class, TradeUtils::getNode);
+        registry.registerItemListing(VillagerTrades.ItemsForEmeralds.class, TradeUtils::getNode);
+        registry.registerItemListing(VillagerTrades.SuspiciousStewForEmerald.class, TradeUtils::getNode);
+        registry.registerItemListing(VillagerTrades.TippedArrowForItemsAndEmeralds.class, TradeUtils::getNode);
+        registry.registerItemListing(VillagerTrades.TreasureMapForEmeralds.class, TradeUtils::getNode);
 
-        registry.registerItemListingCollector(VillagerTrades.DyedArmorForEmeralds.class, DyedArmorForEmeraldsNode::collectItems);
-        registry.registerItemListingCollector(VillagerTrades.EmeraldForItems.class, EmeraldForItemsNode::collectItems);
+        registry.registerItemListingCollector(VillagerTrades.DyedArmorForEmeralds.class, TradeUtils::collectItems);
+        registry.registerItemListingCollector(VillagerTrades.EmeraldForItems.class, TradeUtils::collectItems);
         registry.registerItemListingCollector(VillagerTrades.EmeraldsForVillagerTypeItem.class, EmeraldsForVillagerTypeItemNode::collectItems);
-        registry.registerItemListingCollector(VillagerTrades.EnchantBookForEmeralds.class, EnchantBookForEmeraldsNode::collectItems);
-        registry.registerItemListingCollector(VillagerTrades.EnchantedItemForEmeralds.class, EnchantedItemForEmeraldsNode::collectItems);
-        registry.registerItemListingCollector(VillagerTrades.ItemsAndEmeraldsToItems.class, ItemsAndEmeraldsToItemsNode::collectItems);
-        registry.registerItemListingCollector(VillagerTrades.ItemsForEmeralds.class, ItemsForEmeraldsNode::collectItems);
-        registry.registerItemListingCollector(VillagerTrades.SuspiciousStewForEmerald.class, SuspiciousStewForEmeraldNode::collectItems);
-        registry.registerItemListingCollector(VillagerTrades.TippedArrowForItemsAndEmeralds.class, TippedArrowForItemsAndEmeraldsNode::collectItems);
-        registry.registerItemListingCollector(VillagerTrades.TreasureMapForEmeralds.class, TreasureMapForEmeraldsNode::collectItems);
+        registry.registerItemListingCollector(VillagerTrades.EnchantBookForEmeralds.class, TradeUtils::collectItems);
+        registry.registerItemListingCollector(VillagerTrades.EnchantedItemForEmeralds.class, TradeUtils::collectItems);
+        registry.registerItemListingCollector(VillagerTrades.ItemsAndEmeraldsToItems.class, TradeUtils::collectItems);
+        registry.registerItemListingCollector(VillagerTrades.ItemsForEmeralds.class, TradeUtils::collectItems);
+        registry.registerItemListingCollector(VillagerTrades.SuspiciousStewForEmerald.class, TradeUtils::collectItems);
+        registry.registerItemListingCollector(VillagerTrades.TippedArrowForItemsAndEmeralds.class, TradeUtils::collectItems);
+        registry.registerItemListingCollector(VillagerTrades.TreasureMapForEmeralds.class, TradeUtils::collectItems);
     }
 
     @NotNull