+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Display merchant offer if it's possible to determine it's content even without mod support
 - Added compatibility for Farmer's Delight
+- Added compatibility for Charm Forked
 
 ## [1.3.1]